feat(service): add driver functions to service
